+++ get_rx_persist/lib/get_rx_persist.dart
@@ -3,6 +3,7 @@ library get_rx_persist;
 import 'package:get_rx_persist/src/provider/get_storage_provider.dart';
 import 'package:get_rx_persist/src/storage_provider.dart';
 
+export 'package:get/get_rx/get_rx.dart';
 export 'package:get_rx_persist/src/get_rx_extension.dart';
 export 'package:get_rx_persist/src/provider/mock_storage_provider.dart';