+++ get_rx_persist/lib/src/provider/mock_storage_provider.dart
@@ -1,5 +1,3 @@
-import 'dart:convert';
-
 import 'package:get_rx_persist/src/provider/memory_storage_provider.dart';
 
 /// MockStorageProvider