+++ get_rx_persist/CHANGELOG.md
@@ -1,3 +1,11 @@
-## 0.0.1
+## 0.2.0
 
-* Support Primitive Value
+* Add tests
+* Add comments
+* Add examples
+* Update Doc
+* Support primitive value, collections and object
+
+## 0.1.0
+
+* Support primitive value